ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
python-dotenv==1.0.1
aiohttp==3.11.11
